+++ src/ship.py
@@ -1,14 +1,17 @@
 import pygame
-
-
-class Ship:
+from pygame.sprite import Sprite
+ 
+class Ship(Sprite):
     """A class to manage the ship."""
 
     def __init__(self, ai_game):
         """Initialize the ship and set its starting position."""
+        super().__init__()
         self.screen = ai_game.screen
+        self.ai_game = ai_game
         self.settings = ai_game.settings
         self.screen_rect = ai_game.screen.get_rect()
+        self.ship_group = pygame.sprite.Group()
 
         # Load the ship image  and get its rect.
         self.image = pygame.image.load('../images/ship.bmp')